requirements.txt
fastapi
uvicorn
murf
pillow
pytesseract
pymupdf
ollama
python-multipart
python-dotenv
openai==0.28
fitz
twilio